fix: use correct query params to search for event
Refs #UPS-5427
